All rippers can now use a byte progress bar, added the ability to resume some file downloads

+++ src/main/java/com/rarchives/ripme/ripper/DownloadFileThread.java
@@ -59,18 +59,25 @@ class DownloadFileThread extends Thread {
         this.cookies = cookies;
     }
 
+
     /**
      * Attempts to download the file. Retries as needed.
      * Notifies observers upon completion/error/warn.
      */
     public void run() {
+        long fileSize = 0;
+        int bytesTotal = 0;
+        int bytesDownloaded = 0;
+        if (saveAs.exists() && observer.tryResumeDownload()) {
+            fileSize = saveAs.length();
+        }
         try {
             observer.stopCheck();
         } catch (IOException e) {
             observer.downloadErrored(url, "Download interrupted");
             return;
         }
-        if (saveAs.exists()) {
+        if (saveAs.exists() && !observer.tryResumeDownload()) {
             if (Utils.getConfigBoolean("file.overwrite", false)) {
                 logger.info("[!] Deleting existing file" + prettySaveAs);
                 saveAs.delete();
@@ -80,7 +87,6 @@ class DownloadFileThread extends Thread {
                 return;
             }
         }
-
         URL urlToDownload = this.url;
         boolean redirected = false;
         int tries = 0; // Number of attempts to download
@@ -114,11 +120,20 @@ class DownloadFileThread extends Thread {
                     cookie += key + "=" + cookies.get(key);
                 }
                 huc.setRequestProperty("Cookie", cookie);
+                if (observer.tryResumeDownload()) {
+                    if (fileSize != 0) {
+                        huc.setRequestProperty("Range", "bytes=" + fileSize + "-");
+                    }
+                }
                 logger.debug("Request properties: " + huc.getRequestProperties());
                 huc.connect();
 
                 int statusCode = huc.getResponseCode();
                 logger.debug("Status code: " + statusCode);
+                if (statusCode != 206 && observer.tryResumeDownload() && saveAs.exists()) {
+                    // TODO find a better way to handle servers that don't support resuming downloads then just erroring out
+                    throw new IOException("Server doesn't support resuming downloads");
+                }
                 if (statusCode  / 100 == 3) { // 3xx Redirect
                     if (!redirected) {
                         // Don't increment retries on the first redirect
@@ -146,6 +161,15 @@ class DownloadFileThread extends Thread {
                     observer.downloadErrored(url, "Imgur image is 404: " + url.toExternalForm());
                     return;
                 }
+
+                // If the ripper is using the bytes progress bar set bytesTotal to huc.getContentLength()
+                if (observer.useByteProgessBar()) {
+                    bytesTotal = huc.getContentLength();
+                    observer.setBytesTotal(bytesTotal);
+                    observer.sendUpdate(STATUS.TOTAL_BYTES, bytesTotal);
+                    logger.debug("Size of file at " + this.url + " = " + bytesTotal + "b");
+                }
+
                 // Save file
                 bis = new BufferedInputStream(huc.getInputStream());
 
@@ -154,9 +178,30 @@ class DownloadFileThread extends Thread {
                     String fileExt = URLConnection.guessContentTypeFromStream(bis).replaceAll("image/", "");
                     saveAs = new File(saveAs.toString() + "." + fileExt);
                 }
-
-                fos = new FileOutputStream(saveAs);
-                IOUtils.copy(bis, fos);
+                // If we're resuming a download we append data to the existing file
+                if (statusCode == 206) {
+                    fos = new FileOutputStream(saveAs, true);
+                } else {
+                    fos = new FileOutputStream(saveAs);
+                }
+                byte[] data = new byte[1024 * 256];
+                int bytesRead;
+                while ( (bytesRead = bis.read(data)) != -1) {
+                    try {
+                        observer.stopCheck();
+                    } catch (IOException e) {
+                        observer.downloadErrored(url, "Download interrupted");
+                        return;
+                    }
+                    fos.write(data, 0, bytesRead);
+                    if (observer.useByteProgessBar()) {
+                        bytesDownloaded += bytesRead;
+                        observer.setBytesCompleted(bytesDownloaded);
+                        observer.sendUpdate(STATUS.COMPLETED_BYTES, bytesDownloaded);
+                    }
+                }
+                bis.close();
+                fos.close();
                 break; // Download successful: break out of infinite loop
             } catch (HttpStatusException hse) {
                 logger.debug("HTTP status exception", hse);
